requirements.txt
requests
pydantic
gdm-concordia
python-dotenv
anthropic